Add files via upload
Complete the function insert_at_end(). Function will receive an address of a variable which is holding the address of starting node of the linked list ,an address of a variable which is holding the address of ending node of the linked list and an integer as input, do create a node for the given integer using the predefined function create_node() and insert the newly created node as last node of the given linked list. Note1: Class and methods for doubly Linked List is predefined, do refer it. 
Sample input 1: 678 543 879 808 391 843 179 981 348 799 -1 
Sample output 1: 799 348 981 179 843 391 808 879 543 678